pyproject.toml
[tool.poetry]
name = "projeto-bloco"
version = "0.1.0"
description = ""
authors = ["MatheusFariasSousa <143126173+MatheusFariasSousa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}

pytest = "^8.3.2"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
passlib = "^1.7.4"
jinja2 = "^3.1.4"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"





[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
